# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Inventario",
        "description": "Documentación de la API con Swagger",
        "contact": {
            "email": "soporte@tuempresa.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/salidas-inventario": {
            "get": {
                "tags": [
                    "SalidasInventario"
                ],
                "summary": "Listar salidas de inventario",
                "description": "Listar todas las salidas de inventario",
                "operationId": "4c675d55a9a77b3d8dbd4242f87f954f",
                "responses": {
                    "200": {
                        "description": "Listado de salidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SalidaInventario"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "SalidasInventario"
                ],
                "summary": "Registrar salida de inventario",
                "description": "Registrar una nueva salida de inventario",
                "operationId": "2a1e1915e0644b97a13b25237aa0e38d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lote_id",
                                    "usuario_id",
                                    "cantidad"
                                ],
                                "properties": {
                                    "lote_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "usuario_id": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "cantidad": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "motivo": {
                                        "type": "string",
                                        "example": "Daño en productos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Salida registrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SalidaInventario"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/salidas-inventario/{id}": {
            "get": {
                "tags": [
                    "SalidasInventario"
                ],
                "summary": "Consultar salida por ID",
                "description": "Obtener una salida por ID",
                "operationId": "127d81ca680f869e5329b83f53fcddfc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Salida encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SalidaInventario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Salida no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "SalidasInventario"
                ],
                "summary": "Actualizar salida de inventario",
                "description": "Actualizar una salida existente",
                "operationId": "8e0ce1111f231e5d4345e7bbcf37ff18",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cantidad": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "motivo": {
                                        "type": "string",
                                        "example": "Ajuste de inventario"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Salida actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SalidaInventario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Salida no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "SalidasInventario"
                ],
                "summary": "Eliminar salida de inventario",
                "description": "Eliminar una salida de inventario",
                "operationId": "a3ffcd82edeb22ca6cbc00ccb5c14e24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Salida eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Salida de inventario eliminada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Salida no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SalidaInventario": {
                "required": [
                    "lote_id",
                    "usuario_id",
                    "cantidad"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "lote_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "usuario_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "cantidad": {
                        "type": "integer",
                        "example": 10
                    },
                    "motivo": {
                        "type": "string",
                        "example": "Consumo interno"
                    },
                    "fecha_movimiento": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-04-01T10:30:00Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}